+++ src/main/java/dev/flight_app/dao/DAO.java
@@ -14,9 +14,9 @@ public interface DAO<ID, E extends Identifiable<ID>> {
         return delete(o.id());
     }
 
-    void save(E e);
+//    void save(E e);
 
     void load();
 
-    boolean saveToFile();
+    boolean save();
 }
