+++ src/main/java/dev/flight_app/database/FlightDB.java
@@ -21,7 +21,7 @@ public class FlightDB implements DB<Integer,Flight> {
 
     @Override
     public Map<Integer, Flight> read() {
-        if(!file.exists()) write(new FlightGenerator().getFlights(), file);
+        if(!file.exists()) write(new FlightGenerator().flightsGen(500), file);
         return read(file);
     }
 }
